requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.12.5
pydantic-settings>=2.7.1
boto3>=1.37.1
botocore>=1.37.1
PyYAML>=6.0.2
python-dotenv>=1.0.1
httpx>=0.28.1

